rag-engine/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
pydantic>=2.5.0
python-multipart>=0.0.6
chromadb>=0.4.0
PyMuPDF>=1.23.0
PyPDF2>=3.0.0
python-docx>=1.0.0
python-pptx>=0.6.21
openpyxl>=3.1.0
beautifulsoup4>=4.12.0
chardet>=5.0.0
